# Performance improvements for serialization