--- a/algebra-core/src/serialize/flags.rs
+++ b/algebra-core/src/serialize/flags.rs
@@ -10,18 +10,22 @@ pub trait Flags: Default + Clone + Copy + Sized {
 pub struct EmptyFlags;
 
 impl Flags for EmptyFlags {
+    #[inline]
     fn u8_bitmask(&self) -> u8 {
         0
     }
 
+    #[inline]
     fn from_u8(_value: u8) -> Self {
         EmptyFlags
     }
 
+    #[inline]
     fn from_u8_remove_flags(_value: &mut u8) -> Self {
         EmptyFlags
     }
 
+    #[inline]
     fn len() -> usize {
         0
     }
@@ -37,10 +41,12 @@ pub enum SWFlags {
 }
 
 impl SWFlags {
+    #[inline]
     pub fn infinity() -> Self {
         SWFlags::Infinity
     }
 
+    #[inline]
     pub fn from_y_sign(is_positive: bool) -> Self {
         if is_positive {
             SWFlags::PositiveY
@@ -49,6 +55,7 @@ impl SWFlags {
         }
     }
 
+    #[inline]
     pub fn is_infinity(&self) -> bool {
         match self {
             SWFlags::Infinity => true,
@@ -56,6 +63,7 @@ impl SWFlags {
         }
     }
 
+    #[inline]
     pub fn is_positive(&self) -> Option<bool> {
         match self {
             SWFlags::Infinity => None,
@@ -66,6 +74,7 @@ impl SWFlags {
 }
 
 impl Default for SWFlags {
+    #[inline]
     fn default() -> Self {
         // NegativeY doesn't change the serialization
         SWFlags::NegativeY
@@ -73,6 +82,7 @@ impl Default for SWFlags {
 }
 
 impl Flags for SWFlags {
+    #[inline]
     fn u8_bitmask(&self) -> u8 {
         let mut mask = 0;
         match self {
@@ -83,6 +93,7 @@ impl Flags for SWFlags {
         mask
     }
 
+    #[inline]
     fn from_u8(value: u8) -> Self {
         let x_sign = (value >> 7) & 1 == 1;
         let is_infinity = (value >> 6) & 1 == 1;
@@ -93,6 +104,7 @@ impl Flags for SWFlags {
         }
     }
 
+    #[inline]
     fn from_u8_remove_flags(value: &mut u8) -> Self {
         let flags = Self::from_u8(*value);
         *value &= 0x3F;
@@ -100,6 +112,7 @@ impl Flags for SWFlags {
     }
 
     /// Number of bits required for these flags.
+    #[inline]
     fn len() -> usize {
         2
     }
@@ -114,6 +127,7 @@ pub enum EdwardsFlags {
 }
 
 impl EdwardsFlags {
+    #[inline]
     pub fn from_y_sign(is_positive: bool) -> Self {
         if is_positive {
             EdwardsFlags::PositiveY
@@ -122,6 +136,7 @@ impl EdwardsFlags {
         }
     }
 
+    #[inline]
     pub fn is_positive(&self) -> bool {
         match self {
             EdwardsFlags::PositiveY => true,
@@ -131,6 +146,7 @@ impl EdwardsFlags {
 }
 
 impl Default for EdwardsFlags {
+    #[inline]
     fn default() -> Self {
         // NegativeY doesn't change the serialization
         EdwardsFlags::NegativeY
@@ -138,6 +154,7 @@ impl Default for EdwardsFlags {
 }
 
 impl Flags for EdwardsFlags {
+    #[inline]
     fn u8_bitmask(&self) -> u8 {
         let mut mask = 0;
         match self {
@@ -147,6 +164,7 @@ impl Flags for EdwardsFlags {
         mask
     }
 
+    #[inline]
     fn from_u8(value: u8) -> Self {
         let x_sign = (value >> 7) & 1 == 1;
         if x_sign {
@@ -156,6 +174,7 @@ impl Flags for EdwardsFlags {
         }
     }
 
+    #[inline]
     fn from_u8_remove_flags(value: &mut u8) -> Self {
         let flags = Self::from_u8(*value);
         *value &= 0x7F;
@@ -163,6 +182,7 @@ impl Flags for EdwardsFlags {
     }
 
     /// Number of bits required for these flags.
+    #[inline]
     fn len() -> usize {
         1
     }