This branch adds several performance improvements to the (de-)serialization.
Also see #130.

It mainly inlines small functions and replaces the serialization of BigIntegers/Field points.
I'm not entirely sure why the previous way of serializing these was faster, but the benchmarks consistently show better results using the previous method.

Here's a benchmark between the old serialization from #73, the current one in master, and the one proposed here:
```
write_g1/#73            time:   [109.25 us 109.60 us 110.06 us]
write_g1/current        time:   [115.23 us 115.64 us 116.10 us]
write_g1/#131           time:   [98.144 us 98.384 us 98.714 us]

write_g1_uncompressed/#73
                        time:   [75.913 us 76.064 us 76.266 us]
write_g1_uncompressed/current
                        time:   [92.986 us 93.353 us 93.839 us]
write_g1_uncompressed/#131
                        time:   [74.795 us 75.238 us 75.804 us]

write_field/#73         time:   [38.366 us 38.665 us 39.018 us]
write_field/current     time:   [49.276 us 49.402 us 49.563 us]
write_field/#131        time:   [36.814 us 37.043 us 37.366 us]
```